requirements.txt
streamlit
nltk
cohere
python-docx
pdfplumber
pytesseract
Pillow
python-dotenv
fuzzywuzzy
python-Levenshtein
